+++ docs/changelog.rst
@@ -13,9 +13,11 @@ Unreleased
     New features
     ~~~~~~~~~~~~
 
-..
-    Bug fixes
-    ~~~~~~~~~
+Bug fixes
+~~~~~~~~~
+
+- Improve error message when SQLite fails to open the library database file due
+  to permissions or uncreated target directories. :bug:`1676`
 
 ..
     For plugin developers